refactor console.error override into own iife
